requirements.txt
streamlit
pytube
vosk
transformers
huggingface_hub
torch
RealESRGAN
tqdm
Pillow
opencv-python
torchaudio